+++ src/main/java/com/example/servingwebcontent/GreetingController.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Map;
 
 @Controller
-public class GreetingController02 {
+public class GreetingController {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
@@ -51,22 +51,5 @@ public class GreetingController02 {
 
         return "greeting";
     }
-
-
-    //中本↓↓↓
-    @RequestMapping("/send01")
-    public String send01(){
-        return "send01";
-    }
-    @PostMapping("/receive01") //次回ここから（send01で入力させたデータがattendandesDBのデータと合ってるか判定したい）
-    public String login(
-            Model m,
-            @RequestParam("id") String name,
-            @RequestParam("pass") String pass
-    ) {
-        m.addAttribute("id",name);
-        m.addAttribute("pass",pass);
-        return "receive01";
-    }
-//追加
+    
 }
